docs/requirements.txt
canonical-sphinx[full]
sphinxcontrib-svg2pdfconverter[CairoSVG]
sphinx-last-updated-by-git
sphinxcontrib-mermaid
distro_info
sphinx-hoverxref
sphinx-prompt
sphinxext-rediraffe
sphinx-sitemap
sphinx-togglebutton
railroad-diagrams
